Add GitHub Actions workflow for building and testing (#24)
The new run-tests workflow will run on all branch pushes and pull
requests and simply runs "make build test". If we add additional
tests or static checks we need to remember to update this make
command or add a new phony goal that depends on all goals we want
to trigger in CI executions.