Upgrade Go to 1.17 and start using testing.T.Setenv (#28)
One of the tests set various environment variables with os.Setenv,
but let's use testing.T.Setenv instead to get the original environment
restored automatically. This required us to bump the Go version to 1.17
in go.mod and a few other places. Go 1.17 introduced a slightly
different format for go.mod (indirect dependencies in a separate
"require" block) so that file was adjusted accordingly.